+++ misc/newhall_prism.R
@@ -1,10 +1,13 @@
 # run jNSMR on prism monthly 30yr normals
 
-# path to PRISM data
-PRISM_PATH <- "~/Geodata/PRISM"
+library(terra)
+library(jNSMR)
 
 # PRISM resolution; either 4km or 800m
 RESOLUTION <- "4km"
+
+# path to PRISM data
+PRISM_PATH <- file.path("~/Geodata/PRISM", RESOLUTION)
 
 bilfile <- list.files(PRISM_PATH, "\\.bil$", recursive = TRUE)
 
@@ -24,18 +27,80 @@ colnames(bil) <- c("bilfile","variable","product","resolution","month")
 bilsub <- subset(bil, resolution == RESOLUTION)
 
 # stack rasters, create matrix of values and XY (longlat) locations as data.frame
-prism_stack <- raster::stack(file.path(PRISM_PATH, bilsub$bilfile))
-names(prism_stack) <- paste0(bilsub$variable, bilsub$month)
-prism_frame <- as.data.frame(raster::rasterToPoints(prism_stack))
+# prism_stack <- raster::stack(file.path(PRISM_PATH, bilsub$bilfile))
+
+# terra object
+prism_rast <- terra::rast(file.path(PRISM_PATH, bilsub$bilfile))
 
 # fix column names for jNSM style batch input format
-monthnames <- month.abb[sapply(names(prism_frame), function(x) as.integer(gsub("([a-z])", "", x)))]
-newnames <- gsub("tmean", "t", names(prism_frame))
-newnames <- gsub("ppt", "p", newnames)
-newnames <- gsub("\\d", "", newnames)
-newnames <- paste0(newnames, ifelse(is.na(monthnames), "", monthnames))
-newnames[1:2] <- c("lonDD","latDD")
-names(prism_frame) <- newnames
+monthnames <- month.abb[as.integer(gsub(".*([0-9]{2})_bil$", "\\1", names(prism_rast)))]
+names(prism_rast) <- paste0(gsub("ppt", "p", gsub("tmean","t", 
+                            gsub("PRISM_(ppt|tmean)_.*[0-9]{2}_bil$", "\\1",
+                            names(prism_rast)))),  
+                            ifelse(is.na(monthnames), "", monthnames))
+
+# add longitude and latitude layers
+cx <- terra::xyFromCell(prism_rast, 1:ncell(prism_rast))
+prism_rast$lonDD <- cx[,1]
+prism_rast$latDD <- cx[,2]
+
+# crop for example
+bdy <- soilDB::fetchSDA_spatial(c("CA067","CA620","CA628", "CA630", "CA649", "CA654", "CA651"), "areasymbol", geom.src = "sapolygon")
+                                #c("CA067","CA620","CA628", "CA630", "CA649", "CA654", "CA651")
+x <- terra::crop(prism_rast, terra::vect(bdy))
+
+# PRISM info
+x$stationName <- 1:ncell(x)
+x$awc <- 200
+x$maatmast <- 1.2
+x$pdType <- "Normal"
+x$pdStartYr <- 1981
+x$pdEndYr <- 2010
+x$cntryCode <- "US"
+
+# boilerplate minimum metadata -- these are not used by the algorithm
+x$netType <- ""
+x$elev <- -9999
+x$stProvCode <- ""
+x$notes <- ""
+x$stationID <-1:ncell(x)
+
+# .data <- x
+
+# S3 newhall_batch.SpatRaster method
+# system.time(res <- jNSMR::newhall_batch(x, nrows = 20, cores = 2))
+system.time(res <- jNSMR::newhall_batch(x, nrows = 20))
+
+# inspect result
+bb <- sf::st_as_sf(as(raster::extent(raster::raster(res)), 'SpatialPolygons'))
+sf::st_crs(bb) <- 4326
+sapoly <- soilDB::SDA_spatialQuery(bb, what = 'sapolygon')
+plot(res$NumCumulativeDaysDryOver5C, main = "# Cumulative Days Dry over 5degC\nw/ SAPOLYGON geometry")
+plot(sf::st_geometry(sf::st_cast(bb, 'MULTILINESTRING')), add = TRUE, lwd = 2, col = "RED")
+plot(sf::st_geometry(sapoly), add = TRUE)
+
+# compare areasymbol-summaries of STR/SMR with distribution of classes
+res2 <- soilDB::SDA_query(sprintf( 
+"SELECT areasymbol, taxtempregime, SUM(comppct_r*muacres)/100 AS acres FROM legend
+   INNER JOIN mapunit ON mapunit.lkey = legend.lkey
+   INNER JOIN component ON component.mukey = mapunit.mukey AND majcompflag = 'Yes'
+   WHERE areasymbol IN %s
+   GROUP BY areasymbol, taxtempregime", 
+soilDB::format_SQL_in_statement(unique(sapoly$areasymbol))))
+
+dstr <- dplyr::group_by(res2, areasymbol) |> 
+  dplyr::filter(!is.na(taxtempregime)) |> 
+  dplyr::arrange(dplyr::desc(acres)) |> 
+  dplyr::slice(1)
+sapoly2 <- dplyr::left_join(sapoly, dstr)
+
+plot(res$NumCumulativeDaysDryOver5C,
+     main = "# Cumulative Days MoistDryOver5C\nw/ SAPOLYGON geometry")
+plot(sf::st_geometry(sapoly2), add=TRUE)
+plot(sapoly2['taxtempregime'])
+
+# data.frame interface
+prism_frame <- as.data.frame(terra::as.points(prism_rast))
 
 # PRISM info
 prism_frame$stationName <- 1:nrow(prism_frame)
@@ -60,8 +125,10 @@ test_set <- prism_frame[round(runif(20000, 1, nrow(prism_frame))),]
 afile <- tempfile()
 write.csv(test_set, file = afile) #"misc/prism_monthly.csv")
 test_set <- read.csv(afile)#"misc/prism_monthly.csv")
-# read batch file, run simulations
-res <- jNSMR::newhall_batch(pathname = afile) #"misc/prism_monthly.csv")
+
+# read batch file(s), run simulations. 
+#  using newhall_batch.character S3 wrapper for vector of batch file paths
+system.time({res <- jNSMR::newhall_batch(afile)}) #"misc/prism_monthly.csv")
 
 newsp <- sf::st_as_sf(cbind(test_set[,c('lonDD','latDD')], res), coords = c('lonDD','latDD'))
 sf::st_crs(newsp) <- sf::st_crs(4326)
@@ -69,24 +136,31 @@ plot(newsp$geometry, col=factor(newsp$TemperatureRegime))
 plot(newsp$geometry, col=factor(newsp$MoistureRegime))
 plot(newsp$geometry, col=factor(newsp$RegimeSubdivision1))
 plot(newsp$geometry, col=factor(newsp$RegimeSubdivision2))
-
 newspsub <- subset(newsp, MoistureRegime %in% c("Xeric","Aridic","Ustic","Udic"))
 legnames <- factor(paste(newspsub$MoistureRegime))
 nlegnames <- unique(levels(legnames))
 nlegcolors <- rev(viridisLite::viridis(length(nlegnames)))
-plot(sf::st_transform(newspsub, sf::st_crs(6350))$geometry,
+plot(sf::st_transform(newspsub, sf::st_crs(5070))$geometry,
      col = nlegcolors[match(legnames, nlegnames)], pch=".", cex=5)
-plot(sf::st_transform(spData::us_states$geometry, sf::st_crs(6350)), add=T)
+plot(sf::st_transform(spData::us_states$geometry, sf::st_crs(5070)), add = TRUE)
 legend("topright", legend = nlegnames, pch=".", fill = nlegcolors)
 
 table(newspsub$RegimeSubdivision1)
 
-newspsub2 <- subset(newsp, MoistureRegime %in% c("Aridic","Xeric"))
+newspsub2 <- subset(newsp, MoistureRegime %in% c("Aridic","Xeric","Ustic"))
 legnames <- factor(paste(newspsub2$RegimeSubdivision1, newspsub2$MoistureRegime))
 nlegnames <- unique(levels(legnames))
 nlegcolors <- viridisLite::viridis(length(nlegnames))
-plot(sf::st_transform(newspsub2, sf::st_crs(6350))$geometry,
-     col = nlegcolors[match(legnames, nlegnames)], pch=".", cex=5)
-plot(sf::st_transform(spData::us_states$geometry, sf::st_crs(6350)), add=T)
+plot(sf::st_transform(newspsub2, sf::st_crs(5070))$geometry,
+     col = nlegcolors[match(legnames, nlegnames)], pch = ".", cex = 5)
+plot(sf::st_transform(spData::us_states$geometry, sf::st_crs(5070)), add = TRUE)
+legend("bottomleft", legend = nlegnames, pch=".", fill = nlegcolors)
+
+mlra <- sf::st_read("E:/Geodata/soils/mlra/mlra_v42.shp")
+mlra <- sf::st_transform(mlra, sf::st_crs(5070))
+
+plot(subset(mlra, LRRSYM == 'C')$geometry)
+plot(sf::st_transform(newspsub2, sf::st_crs(5070))$geometry,
+     col = nlegcolors[match(legnames, nlegnames)], pch = ".", cex = 5, add=TRUE)
 legend("bottomleft", legend = nlegnames, pch=".", fill = nlegcolors)