+++ NEWS.md
@@ -1,3 +1,19 @@
+# jNSMR 0.0.0.9007
+
+ - Add SpatRaster and RasterBrick `newhall_batch()` methods for raster input 
+ 
+ - Updated PRISM demo (using SpatRaster interface)
+ 
+# jNSMR 0.0.0.9006
+ 
+ - Added PRISM demo (using data.frame interface) for `newhall_batch()`
+
+# jNSMR 0.0.0.9005
+
+ - Deprecated the JDOM API (JAR v1.6.1 provides legacy support for XML file formats etc.)
+ 
+ - New JAR file v1.6.2 supports command-line inputs for running models
+
 # jNSMR 0.0.0.9004
 
  - Major improvements to handling of metadata and minimum datasets